pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "deduce_uces"
authors = [
    {name = "Cadel Watson", email = "cadel@cadelwatson.com"},
]
readme = "README.md"
classifiers = [ "License :: OSI Approved :: MIT License",]
requires-python = ">=3.7"
dependencies = [
    "biopython>=1.78",
    "psutil>=5.8.0",
    "pysam>=0.16.0.1",
    "networkx>=2.4",
    "scipy==1.7.0",
    "marisa-trie==0.7.7",
    "pandas==1.3.3",
    "numpy==1.21.2",
    "pyfaidx==0.6.2",
    "distributed==2021.11.1",
    "dask==2021.11.1",
    "hypothesis==6.29.0"
]
dynamic = ['version', 'description']

[project.scripts]
deduce = "deduce_uces:main"

[project.optional-dependencies]
test = [
    "tox >= 3.23.0",
    "pytest >=2.7.3",
]
typecheck = [
    "mypy==0.812"
]
